refactor: 지역 관련 DTO 패키지 변경
